envs_requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
albumentations=1.3.1=pypi_0
attrs=23.2.0=pypi_0
bcrypt=4.0.1=pypi_0
beautifulsoup4=4.12.3=pypi_0
better-exceptions=0.3.3=pypi_0
bs4=0.0.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.05.30=haa95532_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.0=pypi_0
cryptography=41.0.3=pypi_0
cycler=0.11.0=pypi_0
einops=0.6.1=pypi_0
et-xmlfile=1.1.0=pypi_0
filelock=3.9.0=pypi_0
fitter=1.6.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.42.1=pypi_0
fsspec=2024.2.0=pypi_0
h11=0.14.0=pypi_0
huggingface-hub=0.21.4=pypi_0
idna=3.4=pypi_0
imageio=2.31.1=pypi_0
imbalanced-learn=0.11.0=pypi_0
imblearn=0.0=pypi_0
itk=5.3.0=pypi_0
itk-core=5.3.0=pypi_0
itk-filtering=5.3.0=pypi_0
itk-io=5.3.0=pypi_0
itk-numerics=5.3.0=pypi_0
itk-registration=5.3.0=pypi_0
itk-segmentation=5.3.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.3.2=pypi_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.2=pypi_0
medpy=0.4.0=pypi_0
mpmath=1.2.1=pypi_0
munch=4.0.0=pypi_0
networkx=3.0=pypi_0
nibabel=5.1.0=pypi_0
numpy=1.26.4=pypi_0
opencv-python=4.8.0.76=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.10=h2bbff1b_1
outcome=1.3.0.post0=pypi_0
packaging=23.1=pypi_0
pandas=2.1.0=pypi_0
paramiko=3.3.1=pypi_0
pillow=9.3.0=pypi_0
pip=23.2.1=py311haa95532_0
pretrainedmodels=0.7.4=pypi_0
protobuf=4.25.3=pypi_0
pycparser=2.21=pypi_0
pydicom=2.4.3=pypi_0
pydot=2.0.0=pypi_0
pynacl=1.5.0=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.0.9=pypi_0
pypdf2=3.0.1=pypi_0
pyreadline3=3.4.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.11.4=he1021f5_0
python-dateutil=2.8.2=pypi_0
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
qudida=0.0.4=pypi_0
requests=2.28.1=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.2=pypi_0
seaborn=0.12.2=pypi_0
selenium=4.21.0=pypi_0
setuptools=68.0.0=py311haa95532_0
simpleitk=2.2.1=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
sympy=1.11.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.8.12=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h2bbff1b_0
torch=2.0.1+cu118=pypi_0
torchaudio=2.0.2+cu118=pypi_0
torchvision=0.15.2+cu118=pypi_0
tqdm=4.66.1=pypi_0
trio=0.25.1=pypi_0
trio-websocket=0.11.1=pypi_0
typing-extensions=4.12.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py311haa95532_0
wsproto=1.2.0=pypi_0
xlrd=1.2.0=pypi_0
xz=5.4.2=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
